Update julia_fun.m
This is the main structure of the code used to produce the figures of 'filled' Julia sets which are in the final report. I used MATLAB to create these figures, and the specific function used in this case (the function can be changed easily) is for the monic polynomial whose filled Julia set is displayed in Figure 11d.